grpc/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.21.12
mysql-connector-python==8.1.0
Flask
flask-cors
requests